+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- OWASP Categories by Severity on the dashboard frontpage
+
 ## [0.0.4] - 2024-01-10
 
 ### Added